refactor separate logic of AccountSettingsDashboard into separate hook
